Add test for passing of Rate-Limit-Token request header
We have some logic to add a `Rate-Limit-Token` request header
when a particular environment variable is present:
https://github.com/alphagov/smokey/blob/22072e261b38da114015a81f8cd73392331d74a5/features/support/env.rb#L49-L58

Prior to this commit, there were no tests in place to prevent the
functionality from being removed or accidentally broken. This is a
quick sense-check Smokey CI test that ensures that the request
header is sent when the `RATE_LIMIT_TOKEN` ENV variable is set.
This frees us up to refactor Smokey's networking with more
confidence. For context on why the functionality was added in the
first place, see:

https://github.com/alphagov/smokey/pull/1048#issuecomment-1302143719