Fix an opaque step name.
The "Check visiting a draft page requires a signon session"
scenario contained the step "When I attempt to go to a case study",
which is super unhelpful because it doesn't say what page it's actually
requesting.

It's also a completely redundant step definition, as it's just a special
case of the "when I visit" step.