Update chromedriver download url, update unzip procedure
Chromedriver distribution has changed - the base URL for chromedriver packages has been updated, and the files that are downloaded have changed structure. They now include an internal directory, and the order is licence file first, then chromedriver.

This means we can't use funzip, since that will only unzip the first file in a zip file. Unzip will handle it more gracefully, but can't take an input from stdin. So we download the file to /tmp, unzip the chromedriver from it, move it to the correct directory, then delete the tmp zipfile.