Remove unhelpful assertions from smoke tests.
These tests contained unnecessary change detectors, which are of negative value
because they break unexpectedly for reasons unrelated to the behaviours
that they're trying to test.

These tests are all meant simply to assert that a user was able to see
the logged-in front page of each app. One of them recently broke because
the string "Sign out" was changed to "Log out".