Add 'govuk_page' abstraction method
We currently have two approaches for querying results in our tests.
One is to use the `page` object provided by Capybara, and another
is to look at the shared `@response` attribute.

The first step in consolidating the two approaches is to provide
a shared method that will handle the complexity. We should always
just be able to refer to the `govuk_page` method, whether or not
we're using `@response` under the hood.