Fix issue with 503'ing assets
Some assets are returning a 503 when going through the proxy to the
mirror. Presumably the assets are missing from the mirror, and it
is AWS/GCP that is returning the 503 as a result.

It's not important for functionality, so let's just stub it.
Fixing the assets issue is outside of the scope of this PR.

```
  Scenario: Check the frontend can talk to Content Store             # features/apps/finder_frontend.feature:7
    When I visit "/government/people"                                # features/step_definitions/smokey_steps.rb:60
    Then I should see "All ministers and senior officials on GOV.UK" # features/step_definitions/smokey_steps.rb:160
    And I should see an input field to search                        # features/step_definitions/finder_frontend_steps.rb:1
      Detected JS errors:

      http://127.0.0.1:8080/assets/static/govuk_publishing_components/icon-close-eecb5fdf67d1688ddf2f93e6e09d5f87b9b24e0fc53596ad2a65cf187c2c7c76.svg - Failed to load resource: the server responded with a status of 503 (Service Unavailable) (RuntimeError)
      ./features/support/hooks.rb:31:in `After'

Failing Scenarios:
cucumber features/apps/finder_frontend.feature:7 # Scenario: Check the frontend can talk to Content Store
```