Added `@worksonmirror` tag to the 17 scenarios that work on mirror
These scenarios pass when tested against the mirror host.
Other scenarios will never pass on the mirror (e.g. scenarios
which rely on search or logging in).

I had to add a `sleep` to allow time for the meta refresh to happen.
Difficult to do much else at the proxy.rb layer, as the
response is a 200, so as far as the proxy is concerned there is
no redirect to wait for, so it feels sensible to handle this at
the step definition layer.

However, we can utilise Capybara under the hood. The `.has_css?`
method will wait until an element matching the selector is found.
`govuk-template` is the class name on the `<html>` element for all
GOV.UK HTML pages, so should be a reliable indicator that the
browser has followed all 301 redirects.