Added `@worksonmirror` tag to the 17 scenarios that work on mirror
These scenarios pass when tested against the mirror host.
Other scenarios will never pass on the mirror (e.g. scenarios
which rely on search or logging in).

Some responses return a 200 even if they are redirects: they
return a blank HTML page with a 'meta fresh' tag which then performs
the redirect. To follow these redirects, there's not much we can do
at the proxy layer, short of adding a `sleep` to allow time for the
refresh to happen. Instead, we utilise Capybara's `has_css?` method
under the hood, to wait until we're on a GOV.UK page - see comment.