Remove old 'Mirrors' feature
This is replaced by the newly tagged up scenarios that can be
run against the mirrors (a much more comprehensive set of
tests).

It also means that the `if @responses` conditional branch of the
`Then I should see` step definition is no longer used, and so has
been deleted here.