Remove the search-analytics smokey tests
The three search queries used in these smokey tests are inflating and
polluting the clickboost search data.

The GA4/Analytics team tried to filter out the Smokey data at the GA4
end but weren't fully able to.

The test can be removed for a few reasons:

1. The clickboost data is being used to evaluate the new search product
and that can't happen effectively if the clickboost data thinks pages
are popular based on clicks occurring in tests. Once the new product is
live it will continue to effect the results that users see.
2. search-analytics hasn't worked properly since the move to kubernetes,
though work is being done to retify this.
3. These tests check for Univeral Analytics which is being removed, so
the tests will need to br written anyhow.