Replace "When I request" with "When I visit", where possible
There's no reason why these particular steps can't use Capybara's
native `visit_path`, as we're not inspecting response codes or
request headers here.

The 200 response code in smokey.feature is implied by the presence
of the text phrase we're looking for in the HTML, so I dropped
that to simplify the smokey.feature test.