Temporarily mark two CDN features as `@notcloudfront`
These should theoretically work on the Cloudfront CDN. However,
we have not made any edits to the `do_http_request` method to
override the host and host headers (as we do in proxy.rb), so the
tests fail.

To avoid complicating an already non-standard Cucumber test, let's
mark this as "not suitable for Cloudfront" for now, and consider
how to refactor our tests further down the line, so that these
two scenarios pass in both the primary and secondary CDN contexts.