Simplify the Smokey run examples
Avoid having to export the `ENVIRONMENT` variable for reuse, as
it's quite clunky. Instead, make use of a Cucumber profile (if it
already exists) and just repeat the environment name (it only
appears twice, and in any case, only production is supported for
the mirror/failoverCDN modes).

We could consider setting up new profiles for the mirror/failover
modes, but that would be part of a separate refactor PR (as we'd
want to avoid duplicating config between cucumber.yml and
govuk_proxy_profiles.rb).