Add tests for GOV.UK Chat
Adds two simple tests:

1. For a non-signed in user viewing a static page. We use the about page
   here instead of the homepage as the Chat app can be turned on and off
   to allow and disallow public access while it's still in beta mode. So
   the homepage can change state, which is why we test against a static
   page that won't change.
2. For a signed-in user viewing the admin area

Chat is set up slightly differently on integration vs staging/prod
currently, as we needed to test some infra work to expose it on the
`/chat` path in integration. So a custom step definition was needed here
to set the public-facing URL differently on integration. Once we've got
all environments consistent, we can delete this whole step definition
file and just use the existing "When I visit x" definition.